backend/requirements.txt
django>=5.0,<7.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=24.0
channels>=4.0
channels-redis>=4.2
daphne>=4.1
celery>=5.3
redis>=5.0
psycopg2-binary>=2.9
python-dotenv>=1.0
requests>=2.31
whitenoise>=6.5
gunicorn>=21.2
openai>=1.0
